+++ packages/api/src/hooks/useSendBoxAttachments.ts
@@ -1,15 +1,17 @@
-import type { WebChatPostActivityAttachment } from 'botframework-webchat-core';
-import { useMemo, type Dispatch, type SetStateAction } from 'react';
+import type { SendBoxAttachment } from 'botframework-webchat-core';
+import { useMemo } from 'react';
 
 import { useSelector } from './internal/WebChatReduxContext';
 import useWebChatAPIContext from './internal/useWebChatAPIContext';
 
 export default function useSendBoxAttachments(): readonly [
-  readonly WebChatPostActivityAttachment[],
-  Dispatch<SetStateAction<readonly WebChatPostActivityAttachment[]>>
+  readonly SendBoxAttachment[],
+  (attachments: readonly SendBoxAttachment[]) => void
 ] {
   // TODO: We should use the selector from "core" package.
-  const sendBoxAttachments = useSelector(({ sendBoxAttachments }) => sendBoxAttachments);
+  const sendBoxAttachments = useSelector(
+    ({ sendBoxAttachments }) => sendBoxAttachments as readonly SendBoxAttachment[]
+  );
   const { setSendBoxAttachments } = useWebChatAPIContext();
 
   return useMemo(