+++ packages/api/src/hooks/useSendMessage.ts
@@ -1,4 +1,4 @@
-import type { WebChatPostActivityAttachment } from 'botframework-webchat-core';
+import type { SendBoxAttachment } from 'botframework-webchat-core';
 import { useCallback } from 'react';
 
 import useWebChatAPIContext from './internal/useWebChatAPIContext';
@@ -7,7 +7,7 @@ import useTrackEvent from './useTrackEvent';
 type SendMessage = (
   text?: string,
   method?: string,
-  init?: { attachments?: readonly WebChatPostActivityAttachment[] | undefined; channelData?: any }
+  init?: { attachments?: readonly SendBoxAttachment[] | undefined; channelData?: any }
 ) => void;
 
 export default function useSendMessage(): SendMessage {
@@ -19,7 +19,7 @@ export default function useSendMessage(): SendMessage {
       trackEvent('sendMessage', {
         numAttachments: attachments?.length || 0,
         // eslint-disable-next-line no-magic-numbers
-        sumSizeInKB: Math.round(attachments?.reduce((total, { size }) => total + size, 0) / 1024)
+        sumSizeInKB: Math.round(attachments?.reduce((total, { blob: { size } }) => total + size, 0) / 1024)
       });
 
       sendMessage(text, method, { attachments, channelData });