+++ packages/component/src/hooks/useSendFiles.ts
@@ -1,7 +1,7 @@
 import { hooks } from 'botframework-webchat-api';
 import { useCallback } from 'react';
 
-import useConvertFileToAttachment from './internal/useConvertFileToAttachment';
+import useMakeThumbnail from './internal/useMakeThumbnail';
 
 const { useSendFiles: useAPISendFiles } = hooks;
 
@@ -9,8 +9,8 @@ const { useSendFiles: useAPISendFiles } = hooks;
  * @deprecated This hook will be removed on or after 2026-04-03. Please use `useSendMessage` instead.
  */
 export default function useSendFiles(): (files: readonly File[]) => void {
+  const makeThumbnail = useMakeThumbnail();
   const sendFiles = useAPISendFiles();
-  const convertFileToAttachment = useConvertFileToAttachment();
 
   return useCallback(
     files => {
@@ -18,9 +18,13 @@ export default function useSendFiles(): (files: readonly File[]) => void {
       // This is the because the Promise returned never tell if the message has successfully sent or not.
       // Until we have that signal, we should not return Promise.
       (async function () {
-        files && files.length && sendFiles(await Promise.all([...files].map(convertFileToAttachment)));
+        files &&
+          files.length &&
+          sendFiles(
+            await Promise.all(files.map(blob => makeThumbnail(blob).then(thumbnailURL => ({ blob, thumbnailURL }))))
+          );
       })();
     },
-    [convertFileToAttachment, sendFiles]
+    [makeThumbnail, sendFiles]
   );
 }