+++ packages/component/src/hooks/useSendMessage.ts
@@ -1,7 +1,7 @@
 import { hooks } from 'botframework-webchat-api';
 import { useCallback } from 'react';
 
-import useConvertFileToAttachment from './internal/useConvertFileToAttachment';
+import useMakeThumbnail from './internal/useMakeThumbnail';
 
 const { useSendMessage: useAPISendMessage } = hooks;
 
@@ -12,16 +12,22 @@ type SendMessage = (
 ) => void;
 
 export default function useSendMessage(): SendMessage {
-  const convertFileToAttachment = useConvertFileToAttachment();
+  const makeThumbnail = useMakeThumbnail();
   const sendMessage = useAPISendMessage();
 
   return useCallback<SendMessage>(
     async (text, method, { channelData, files } = {}) => {
       sendMessage(text, method, {
-        attachments: files ? await Promise.all([...files].map(convertFileToAttachment)) : [],
+        attachments: files
+          ? await Promise.all(
+              await Promise.all(
+                [...files].map(blob => makeThumbnail(blob).then(thumbnailURL => ({ blob, thumbnailURL })))
+              )
+            )
+          : [],
         channelData
       });
     },
-    [convertFileToAttachment, sendMessage]
+    [makeThumbnail, sendMessage]
   );
 }