+++ packages/core/src/actions/sendFiles.ts
@@ -1,9 +1,14 @@
-import type { WebChatPostActivityAttachment } from '../types/WebChatPostActivityAttachment';
-
 const SEND_FILES = 'WEB_CHAT/SEND_FILES';
 
 // TODO: [P1] We could obsolete this or dispatch { type: SEND_MESSAGE } insetad.
-export default function sendFiles(files: readonly WebChatPostActivityAttachment[]) {
+export default function sendFiles(
+  files: readonly Readonly<{
+    name: string;
+    size: number;
+    url: string;
+    thumbnail?: string;
+  }>[]
+) {
   return {
     type: SEND_FILES,
     payload: { files }