+++ packages/core/src/actions/setSendBoxAttachments.ts
@@ -1,6 +1,8 @@
+import type { SendBoxAttachment } from '../types/SendBoxAttachment';
+
 const SET_SEND_BOX_ATTACHMENTS = 'WEB_CHAT/SET_SEND_BOX_ATTACHMENTS';
 
-export default function setSendBoxAttachments(attachments: readonly Blob[]) {
+export default function setSendBoxAttachments(attachments: readonly SendBoxAttachment[]) {
   return {
     type: SET_SEND_BOX_ATTACHMENTS,
     payload: { attachments }