+++ packages/core/src/reducers/sendBoxAttachments.ts
@@ -1,8 +1,9 @@
 import { SET_SEND_BOX_ATTACHMENTS } from '../actions/setSendBoxAttachments';
+import type { SendBoxAttachment } from '../types/SendBoxAttachment';
 
-const DEFAULT_STATE: readonly Blob[] = Object.freeze([]);
+const DEFAULT_STATE: readonly SendBoxAttachment[] = Object.freeze([]);
 
-export default function sendBoxAttachments(state = DEFAULT_STATE, { payload, type }): readonly Blob[] {
+export default function sendBoxAttachments(state = DEFAULT_STATE, { payload, type }): readonly SendBoxAttachment[] {
   switch (type) {
     case SET_SEND_BOX_ATTACHMENTS:
       state = payload.attachments;