+++ packages/core/src/selectors/sendBoxAttachments.ts
@@ -1,5 +1,6 @@
+import type { SendBoxAttachment } from '../types/SendBoxAttachment';
 import type { ReduxState } from '../types/internal/ReduxState';
 
-export default function sendBoxAttachments({ sendBoxAttachments }: ReduxState): readonly Blob[] {
+export default function sendBoxAttachments({ sendBoxAttachments }: ReduxState): readonly SendBoxAttachment[] {
   return sendBoxAttachments;
 }