+++ packages/core/src/types/SendBoxAttachment.ts
@@ -1,0 +1,4 @@
+export type SendBoxAttachment = Readonly<{
+  blob: Blob | File;
+  thumbnailURL?: URL;
+}>;